requirements/testing.txt
-r ./core.txt
docutils
mock
packaging
poethepoet
PyHamcrest
Pygments
pytest
